move the functionlity of trimFileName functiolity to lib/file.ts
